remove unused components
