+++ docs/flashbots-mev-share/searchers/getting-started.mdx
@@ -2,8 +2,8 @@
 title: Getting Started
 ---
 
-import Tabs from '@theme/Tabs'
-import TabItem from '@theme/TabItem'
+import Tabs from '@theme/Tabs';
+import TabItem from '@theme/TabItem';
 
 To start searching on MEV-Share, you will first need to connect to a MEV-Share Node. As a reminder, the MEV-Share Node is responsible for receiving transactions and bundles from users, and selectively sharing information ("hints") about them with searchers. When a searcher wants to include a transaction in their bundle, they use that transaction's hash. The MEV-Share Node replaces it with the original transaction before sending the bundle to a block builder.
 
@@ -16,44 +16,44 @@ The simplest way to connect to the Flashbots MEV-Share Node is to use a [client 
 <Tabs>
 <TabItem value="ts" label="mev-share-client-ts">
 
-*Add library to your project:*
+_Add library to your project:_
 
 ```bash
 yarn add @flashbots/mev-share-client
 ```
 
-*Use the following code to import the library* (Replace ALL_CAPS placeholders with your data):
-```typescript
-import { Wallet, JsonRpcProvider } from "ethers"
-import MevShareClient, {
-    BundleParams,
-    HintPreferences,
-    IPendingBundle,
-    IPendingTransaction,
-    TransactionOptions
-} from "@flashbots/mev-share-client"
+_Use the following code to import the library_ (Replace ALL_CAPS placeholders with your data):
 
-const provider = new JsonRpcProvider(RPC_URL)
-const authSigner = new Wallet(FB_REPUTATION_PRIVATE_KEY, provider)
-const mevShareClient = MevShareClient.useEthereumMainnet(authSigner)
+```typescript
+import {Wallet, JsonRpcProvider} from 'ethers';
+import MevShareClient, {
+  BundleParams,
+  IPendingBundle,
+  IPendingTransaction,
+  TransactionOptions,
+} from '@flashbots/mev-share-client';
+
+const provider = new JsonRpcProvider(RPC_URL);
+const authSigner = new Wallet(FB_REPUTATION_PRIVATE_KEY, provider);
+const mevShareClient = MevShareClient.useEthereumMainnet(authSigner);
 ```
 
 _Connecting to Goerli:_
 
 ```typescript
-const mevShareClient = MevShareClient.useEthereumGoerli(authSigner)
+const mevShareClient = MevShareClient.useEthereumGoerli(authSigner);
 ```
 
-*Advanced setup* (for developers):
+_Advanced setup_ (for developers):
 
 ```typescript
 // connect to MEV-Share on mainnet
 const mevShareClient = new MevShareClient(authSigner, {
-    name: "mainnet",
-    chainId: 1,
-    streamUrl: "https://mev-share.flashbots.net",
-    apiUrl: "https://relay.flashbots.net",
-})
+  name: 'mainnet',
+  chainId: 1,
+  streamUrl: 'https://mev-share.flashbots.net',
+  apiUrl: 'https://relay.flashbots.net',
+});
 ```
 
 Further documentation on the client library can be found in the [mev-share-client-ts](https://github.com/flashbots/mev-share-client-ts).
@@ -74,9 +74,9 @@ Now you should be connected to the Flashbots MEV-Share Node. Continue reading on
 Searching on MEV-Share is different from searching on the mempool in that only certain parts of a transaction are shared with searchers. In the mempool, we can see all parts of a transaction, such as its calldata or who the transaction is from. But on MEV-Share, a transaction might only reveal its function selector, making a traditional arbitrage calculation infeasible.
 
 There are three primary strategies for searching on MEV-Share:
-* Probabilistically: send many bundles that probabilistically backrun MEV-Share orderflow.
-* On-chain: perform more of your searching on-chain instead of off-chain.
-* Existing: only search on transactions which share all the information you need.
 
-To maximally leverage MEV-Share searchers will need to employ new strategies.
-:::
+- Probabilistically: send many bundles that probabilistically backrun MEV-Share orderflow.
+- On-chain: perform more of your searching on-chain instead of off-chain.
+- Existing: only search on transactions which share all the information you need.
+
+To maximally leverage MEV-Share searchers will need to employ new strategies. :::
