+++ docs/flashbots-protect/mev-share.mdx
@@ -4,7 +4,7 @@ title: MEV-Share
 
 import HintsTable from '../specs/mev-share/HintsTable';
 import Builders from '../specs/mev-share/_builders.mdx';
-import ProtectButtonSelector from '@site/src/components/ProtectButtonSelector';
+import ProtectButton from '@site/src/components/ProtectButton';
 
 ## Introduction
 
@@ -14,7 +14,7 @@ MEV-Share enables users to reclaim up to 90% of the MEV generated by their trans
 
 Advanced users can exert more control over their transactions and preferences through the advanced panel or by manually configuring their Protect RPC request.
 
-<ProtectButtonSelector />
+<ProtectButton />
 
 ## Common configurations
 
@@ -116,6 +116,4 @@ All percentages in the refund parameters must total less than 100. The remaining
 
 ### Priority fee
 
-When sending private transaction to rpc endpoint you should set priority fee (tips) to be strictly greater than zero.
-Transactions with 0 priority fee will not be shared with block builders and included on chain, unless they are bundled
-by a searcher via MEV-Share.
+When sending private transaction to rpc endpoint you should set priority fee (tips) to be strictly greater than zero. Transactions with 0 priority fee will not be shared with block builders and included on chain, unless they are bundled by a searcher via MEV-Share.
